Removed type CustomPost
